Improvements and fixes to fetchdata, inputwindow and database. The addition of the new columns in the games database needed to be taken into account when storing games in the database. This fixes that issue.
Bugfix for _tryAltertatives() function, which was checking the wrong url so we didn't get the release info from it.

Add input fields for publisher and developer to inputwindow, which will be autofilled by fetchinfo and used internally for the future side panel. A limitation of this currently is that the publisher will always be the same no matter the region since that data comes from the main data part of MG rather than the release info part.

Add possibility to search for a specific country's release.

Add fetchinfo.printInfo() so we keep it in one place.

Add inputwindow._displayMsgBox to make it a bit more DRY.