Add sidepanel core functionality SidePanel implements a details side panel with basic but working functionality.
MainWindow: As a result of this I also got rid of the grid layout in the main window (because grids are confusing...) and use a combination of QHBox and QVBox layouts instead.
Also add a new context menu button for opening the details side-panel.

Table(): Override keyPressEvent (to intercept the return key and start editing instead of doing nothing) and mouseDoubleClickEvent (to make double clicks run the rowData() method which sends data to the side panel) methods. This currently only works with the games table since the side panel needs to be set up properly first.

InputWindow: Attempt at fixing issue with adding games that have single quotes in their name, but it needs to be fixed in the SQL statements in Table.addData() as well.