Table improvements and other fixes Table: Rewrite Table.addData() to use QSqlTableModel.insertRecord() instead of using direct queries with QSqlQuery. This is both cleaner and less error prone, and also fixes the issue with adding games with single-quotes in their columns as a side effect.
Move TableModel into table.py and delete models.py since it was really only the TableModel in there that was used, and it can reside in the same file as the Table class itself.

Don't save the QSqlDatabase object as a member variable anywhere anymore. This didn't break anything currently but can result in undefined behavior on application exit. The application is now getting the db by using the QSqlTableModel.database() method instead where it was used before.

createDB: Change the sqlite3 creation function to use INTEGER PRIMARY KEY for the ID column. This was probably unnecessary, but was done in an attempt to fix an issue with the above insertRecord change. The issue was not related, but this change doesn't really hurt and it's good to be explicit. Also updated the included collection.db to use this.

sql2csv: Add the new columns so they get exported properly as well.

getMobyRelease: Add "EAN-13" or "UPC-A" to the value of the "code" key so the user knows what the code is, in case the function doesn't find any other product code.

InputWindow: Fix typo.