Accurate rowcount results for search text Table: Make filterTable() return the number of items in a search. This slowed down the search speed somewhat unfortunately. Especially noticable for filters that have a lot of results.
MainWindow: Use the filterTable return value to update the statusbar with an accurate number. This makes it accurate for results that have >256 entries as well.
Also because of the slower search, make the searchBar only do a search when user presses Return. Might look into changing it back when we have asyncio functionality.