Update VGDB dats to work with fetchinfo All the VGDB dats used region names like "North America" and "Europe" instead of "NTSC (NA)" and "PAL". Replace all instances of those to the correct ones, and fix some other inconsistencies in a few of them.
Also add platform substitutions in getMobyInfo() because the platform names imported from the VGDB dats differ from what mobygames expects.
And temporarily comment out the logic that skips the "Package Comments" cover images because it didn't work reliably and in some cases it skipped valid covers (such as for FDS games where the comment would be 'FDS Release' or similar).