Tweak size restriction on cover label. With a max width of 250 it made cover images be not centered in the side panel. By changing it to the width of the side panel itself this is fixed.
Also set minimum height to 200. This seems to be just enough to where even games that are available for ALL THE PLATFORMS (like Bubble Bobble) can display most of the image without overflowing into the details text.

Remove the coverLabel from the layout because it's obviously a cover. And use that gained space by adding a fancy horizontal line underneath the cover while we're at it.